refactor(svelte): split pure semantic-keys from rune-backed service
Move source-identity trackers, epoch tokens, and resolveSemanticKeys into
semantic-keys.ts (no runes). Keep createSemanticKeyService in
semantic-keys.svelte.ts. Split co-located pure vs service characterization
tests and point production/test imports at the owning module.
